pyproject.toml
[project]
name = "proyecto"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "langchain>=0.3.1",
    "langchain-google-genai>=2.0.8",
    "langchain-openai>=0.3.1",
    "python-dotenv>=1.0.1",
    "requests>=2.33.1",
    "selenium>=4.43.0",
    "streamlit>=1.32.0",
    "trafilatura>=2.0.0",
]
